// frontend/src/components/radio/SpotifyPlayer/types.ts
export type TrackFeature = {
  active: boolean;
  value: number;
};

export type TrackFeatures = {
  key: TrackFeature;
  mode: TrackFeature;
  tempo: TrackFeature;
  danceability: TrackFeature;
  energy: TrackFeature;
  loudness: TrackFeature;
  speechiness: TrackFeature;
  acousticness: TrackFeature;
  instrumentalness: TrackFeature;
  liveness: TrackFeature;
  valence: TrackFeature;
};